# openapi.yml
openapi: 3.0.3
info:
  title: Hyper Fleet Native API
  description: |
    Hyper Fleet Native API provides comprehensive management of virtual machines (sandboxes), 
    file system operations, execution environments, and organizational resources.

    All endpoints except `/api/register` require the `X-API-Key` header for authentication.
  version: 1.0.0
  contact:
    name: Hyper Fleet Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:33944/api
    description: Local development server
  - url: https://api.hyperfleet.io/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for authentication (format: hf_...)"
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Response data (structure varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        details:
          type: string
          example: Additional error details

    # Registration
...
    ExecResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: ok
        data:
          type: object
          properties:
            stdout:
              type: string
              example: |
                total 16
                drwx------  2 root root 4096 Jan  1 12:00 .
                drwxr-xr-x 18 root root 4096 Jan  1 12:00 ..
            stderr:
              type: string
            exitCode:
              type: integer
              example: 0

    # Background Process Management
...
    FileInfo:
      type: object
      properties:
        name:
          type: string
          example: file.txt
        path:
          type: string
          example: /root/file.txt
        size:
          type: integer
          example: 1024
        mode:
          type: string
          example: 0644
        isDir:
          type: boolean
          example: false
        modTime:
          type: string
          format: date-time
...
    CreateFileRequest:
      type: object
      properties:
        content:
          type: string
          example: hello world
...
paths:
...
  /sandboxes/{id}/files:
    get:
      tags:
        - File System
      summary: List files
      description: List files and directories in a path
      operationId: listFiles
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /root
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      files:
                        type: array
                        items:
                          $ref: "#/components/schemas/FileInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Sandbox or path not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - File System
      summary: Delete file or directory
      description: Delete a file or directory
      operationId: deleteFile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /root/file.txt
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/upload:
    post:
      tags:
        - File System
      summary: Upload file
      description: Upload a file to the sandbox
      operationId: uploadFile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /root/script.sh
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/download:
    get:
      tags:
        - File System
      summary: Download file
      description: Download a file from the sandbox
      operationId: downloadFile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /root/output.log
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/head-tail:
    get:
      tags:
        - File System
      summary: Read file head or tail
      description: Get the first or last N lines of a file
      operationId: headTail
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /var/log/syslog
        - name: lines
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
          example: 50
        - name: head
          in: query
          schema:
            type: boolean
            default: true
          description: If true, return head (first lines), if false, return tail (last lines)
          example: false
      responses:
        "200":
          description: File content (head or tail)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/search:
    get:
      tags:
        - File System
      summary: Search files
      description: Search for files matching a pattern
      operationId: searchFiles
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /var/log
        - name: pattern
          in: query
          required: true
          schema:
            type: string
          example: "*.error"
      responses:
        "200":
          description: Matching files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Path not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/mkdir:
    post:
      tags:
        - File System
      summary: Create directory
      description: Create a new directory
      operationId: createDirectory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /new/dir
      responses:
        "200":
          description: Directory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid path or already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/create:
    post:
      tags:
        - File System
      summary: Create file
      description: Create a new file with optional content
      operationId: createFile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: path
          in: query
          required: true
          schema:
            type: string
          example: /file.txt
        - name: content
          in: query
          schema:
            type: string
          example: hello
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileRequest"
      responses:
        "200":
          description: File created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/copy:
    post:
      tags:
        - File System
      summary: Copy file or directory
      description: Copy a file or directory to a new location
      operationId: copyFile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 65ae1234567890abcdef1234
        - name: from
          in: query
          required: true
          schema:
            type: string
          example: /src
        - name: to
          in: query
          required: true
          schema:
            type: string
          example: /dst
      responses:
        "200":
          description: File copied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid paths or source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...